sha-crypt: Change `rounds` parameter type to `u32`
Its type doesn't need to depend on the target architecture pointer
width.

Fixes #498.

+++ sha-crypt/tests/lib.rs
@@ -10,7 +10,7 @@ struct TestVector {
     salt: &'static str,
     result_sha256: &'static str,
     result_sha512: &'static str,
-    rounds: usize,
+    rounds: u32,
 }
 
 const TEST_VECTORS: &[TestVector] = &[
